Implementa a Dto,Entity e UseCase de Order e Metodos de Segurança

+++ src/main/java/com/Ebooks/Ebooks_api/Entity/Book.java
@@ -16,7 +16,7 @@ public class Book {
     private String Nome;
     private String img;
     private String autor;
-    private Integer preco;
+    private Long preco;
     private Integer score;
 
     @Column(name = "user_id")
